+++ src/assets/icons/icon-cardano.tsx
@@ -1,22 +1,21 @@
 import React from 'react';
 import Svg, { Path } from 'react-native-svg';
 
-import { themes } from '@/themes';
-
-import { IconProps } from './types';
+import { colors } from '@/helpers/themes';
+import { IconProps } from '@/types/IconType';
 
 export function IconCardano({ size = 25 }: Omit<IconProps, 'color'>) {
 	return (
 		<Svg width={size} height={size} viewBox="0 0 72 72" fill="none">
 			<Path
 				d="M36 72C55.8823 72 72 55.8823 72 36C72 16.1177 55.8823 0 36 0C16.1177 0 0 16.1177 0 36C0 55.8823 16.1177 72 36 72Z"
-				fill={themes.colors.dark_grey}
+				fill={colors.dark_grey}
 			/>
 			<Path
 				fill-rule="evenodd"
 				clip-rule="evenodd"
 				d="M31.9857 24.9567C32.6895 24.9568 33.3774 25.1656 33.9625 25.5567C34.5476 25.9478 35.0036 26.5035 35.2728 27.1538C35.5421 27.804 35.6124 28.5194 35.4751 29.2096C35.3377 29.8999 34.9988 30.5339 34.5011 31.0315C34.0034 31.529 33.3694 31.8679 32.6791 32.0051C31.9889 32.1424 31.2735 32.0719 30.6233 31.8026C29.9731 31.5332 29.4174 31.0771 29.0264 30.492C28.6355 29.9068 28.4268 29.2189 28.4268 28.5151C28.4269 27.5713 28.802 26.6662 29.4694 25.9989C30.1368 25.3316 31.0419 24.9567 31.9857 24.9567ZM40.5768 24.9567C41.2805 24.9569 41.9684 25.1657 42.5535 25.5568C43.1385 25.9479 43.5945 26.5037 43.8637 27.1539C44.1329 27.8042 44.2032 28.5196 44.0658 29.2098C43.9284 29.9 43.5895 30.534 43.0918 31.0315C42.5941 31.5291 41.9601 31.8679 41.2698 32.0052C40.5796 32.1424 39.8642 32.0719 39.214 31.8025C38.5638 31.5332 38.0082 31.0771 37.6172 30.4919C37.2262 29.9068 37.0176 29.2188 37.0176 28.5151C37.0177 27.5713 37.3928 26.6661 38.0602 25.9988C38.7277 25.3315 39.6329 24.9566 40.5768 24.9567ZM40.5768 39.9265C41.2805 39.9267 41.9684 40.1355 42.5535 40.5266C43.1385 40.9177 43.5945 41.4735 43.8637 42.1238C44.1329 42.774 44.2032 43.4894 44.0658 44.1796C43.9284 44.8698 43.5895 45.5038 43.0918 46.0014C42.5941 46.4989 41.9601 46.8377 41.2698 46.975C40.5796 47.1122 39.8642 47.0417 39.214 46.7723C38.5638 46.503 38.0082 46.0469 37.6172 45.4618C37.2262 44.8766 37.0176 44.1886 37.0176 43.4849C37.0176 43.0175 37.1097 42.5548 37.2886 42.123C37.4675 41.6913 37.7297 41.299 38.0602 40.9685C38.3907 40.6381 38.783 40.376 39.2148 40.1972C39.6466 40.0184 40.1094 39.9265 40.5768 39.9265ZM31.9857 39.9265C32.6895 39.9266 33.3774 40.1354 33.9625 40.5265C34.5476 40.9176 35.0036 41.4734 35.2728 42.1236C35.5421 42.7738 35.6124 43.4892 35.4751 44.1795C35.3377 44.8697 34.9988 45.5037 34.5011 46.0013C34.0034 46.4988 33.3694 46.8377 32.6791 46.9749C31.9889 47.1122 31.2735 47.0417 30.6233 46.7724C29.9731 46.503 29.4174 46.0469 29.0264 45.4618C28.6355 44.8766 28.4268 44.1887 28.4268 43.4849C28.4268 43.0176 28.5189 42.5548 28.6978 42.1231C28.8766 41.6914 29.1388 41.2991 29.4693 40.9686C29.7998 40.6382 30.1921 40.3761 30.6239 40.1973C31.0556 40.0185 31.5184 39.9265 31.9857 39.9265ZM27.5676 32.5642C28.2713 32.5644 28.9592 32.7732 29.5443 33.1642C30.1294 33.5553 30.5854 34.1111 30.8547 34.7613C31.1239 35.4115 31.1943 36.127 31.0569 36.8172C30.9196 37.5074 30.5806 38.1414 30.083 38.639C29.5853 39.1366 28.9512 39.4754 28.261 39.6127C27.5708 39.7499 26.8553 39.6794 26.2051 39.4101C25.555 39.1407 24.9993 38.6846 24.6083 38.0995C24.2173 37.5143 24.0086 36.8264 24.0086 36.1226C24.0088 35.1788 24.3838 34.2737 25.0512 33.6064C25.7186 32.9391 26.6238 32.5642 27.5676 32.5642ZM44.9949 32.5642C45.6987 32.5644 46.3866 32.7733 46.9716 33.1644C47.5567 33.5555 48.0127 34.1113 48.2818 34.7615C48.551 35.4117 48.6214 36.1271 48.484 36.8173C48.3466 37.5075 48.0076 38.1415 47.5099 38.6391C47.0123 39.1366 46.3782 39.4755 45.688 39.6127C44.9977 39.7499 44.2823 39.6794 43.6322 39.4101C42.982 39.1407 42.4263 38.6846 42.0354 38.0995C41.6444 37.5143 41.4357 36.8264 41.4357 36.1226C41.4358 35.6553 41.5278 35.1925 41.7067 34.7607C41.8856 34.329 42.1478 33.9367 42.4783 33.6063C42.8088 33.2758 43.2012 33.0137 43.633 32.8349C44.0648 32.6562 44.5276 32.5642 44.9949 32.5642ZM49.0449 26.4293C49.4817 26.4295 49.9087 26.5591 50.2719 26.8019C50.635 27.0446 50.918 27.3896 51.0851 27.7932C51.2522 28.1968 51.2958 28.6408 51.2105 29.0692C51.1252 29.4976 50.9148 29.8911 50.6059 30.2C50.297 30.5088 49.9035 30.7191 49.4751 30.8043C49.0466 30.8894 48.6026 30.8457 48.199 30.6785C47.7955 30.5113 47.4506 30.2282 47.2079 29.865C46.9652 29.5018 46.8357 29.0748 46.8357 28.638C46.8357 28.3479 46.8929 28.0606 47.0039 27.7926C47.1149 27.5246 47.2777 27.2811 47.4828 27.076C47.688 26.8709 47.9315 26.7083 48.1995 26.5973C48.4676 26.4863 48.7548 26.4293 49.0449 26.4293ZM49.0449 41.1533C49.4817 41.1535 49.9087 41.2831 50.2719 41.5258C50.635 41.7686 50.918 42.1136 51.0851 42.5172C51.2522 42.9208 51.2958 43.3648 51.2105 43.7932C51.1252 44.2216 50.9148 44.6151 50.6059 44.924C50.297 45.2328 49.9035 45.4431 49.4751 45.5283C49.0466 45.6134 48.6026 45.5697 48.199 45.4025C47.7955 45.2353 47.4506 44.9522 47.2079 44.589C46.9652 44.2258 46.8357 43.7988 46.8357 43.362C46.8357 43.0719 46.8929 42.7847 47.0039 42.5167C47.1149 42.2487 47.2777 42.0051 47.4828 41.8C47.688 41.5949 47.9315 41.4323 48.1995 41.3213C48.4676 41.2104 48.7548 41.1533 49.0449 41.1533ZM23.5176 41.1533C23.9544 41.1534 24.3814 41.283 24.7445 41.5257C25.1077 41.7684 25.3908 42.1134 25.5579 42.517C25.725 42.9206 25.7687 43.3646 25.6835 43.7931C25.5982 44.2215 25.3878 44.615 25.0789 44.9239C24.77 45.2327 24.3765 45.443 23.9481 45.5282C23.5196 45.6134 23.0756 45.5697 22.672 45.4025C22.2684 45.2353 21.9235 44.9522 21.6808 44.589C21.4382 44.2258 21.3086 43.7988 21.3086 43.362C21.3087 42.7762 21.5415 42.2144 21.9557 41.8002C22.37 41.386 22.9318 41.1533 23.5176 41.1533ZM23.5176 26.4293C23.9544 26.4294 24.3814 26.559 24.7445 26.8017C25.1077 27.0444 25.3908 27.3894 25.5579 27.793C25.725 28.1966 25.7687 28.6406 25.6835 29.0691C25.5982 29.4975 25.3878 29.891 25.0789 30.1999C24.77 30.5087 24.3765 30.719 23.9481 30.8042C23.5196 30.8894 23.0756 30.8457 22.672 30.6785C22.2684 30.5113 21.9235 30.2282 21.6808 29.865C21.4382 29.5018 21.3086 29.0748 21.3086 28.638C21.3087 28.0522 21.5415 27.4904 21.9557 27.0762C22.37 26.662 22.9318 26.4293 23.5176 26.4293ZM36.2812 19.0671C36.7181 19.0672 37.145 19.1968 37.5082 19.4396C37.8713 19.6823 38.1543 20.0273 38.3214 20.4309C38.4885 20.8345 38.5321 21.2786 38.4468 21.707C38.3616 22.1354 38.1512 22.5288 37.8422 22.8377C37.5333 23.1465 37.1398 23.3568 36.7114 23.442C36.2829 23.5272 35.8389 23.4834 35.4353 23.3162C35.0318 23.149 34.6869 22.8659 34.4442 22.5027C34.2015 22.1395 34.072 21.7125 34.072 21.2757C34.072 20.9856 34.1292 20.6984 34.2402 20.4304C34.3512 20.1624 34.514 19.9189 34.7191 19.7138C34.9243 19.5087 35.1678 19.346 35.4359 19.235C35.7039 19.1241 35.9912 19.067 36.2812 19.0671ZM36.2812 48.5156C36.7181 48.5157 37.145 48.6454 37.5082 48.8881C37.8713 49.1309 38.1543 49.4759 38.3214 49.8795C38.4885 50.283 38.5321 50.7271 38.4468 51.1555C38.3616 51.5839 38.1512 51.9774 37.8422 52.2862C37.5333 52.5951 37.1398 52.8054 36.7114 52.8905C36.2829 52.9757 35.8389 52.9319 35.4353 52.7648C35.0318 52.5976 34.6869 52.3145 34.4442 51.9513C34.2015 51.5881 34.072 51.1611 34.072 50.7243C34.072 50.4342 34.1292 50.1469 34.2402 49.8789C34.3512 49.6109 34.514 49.3674 34.7191 49.1623C34.9243 48.9572 35.1678 48.7946 35.4359 48.6836C35.7039 48.5726 35.9912 48.5156 36.2812 48.5156ZM45.7313 50.7243C46.0952 50.7244 46.451 50.8324 46.7536 51.0347C47.0562 51.237 47.292 51.5245 47.4312 51.8608C47.5705 52.1972 47.6068 52.5672 47.5358 52.9242C47.4647 53.2812 47.2893 53.6091 47.0319 53.8664C46.7745 54.1238 46.4465 54.299 46.0895 54.37C45.7325 54.4409 45.3625 54.4044 45.0262 54.2651C44.69 54.1258 44.4025 53.8899 44.2003 53.5873C43.9981 53.2846 43.8902 52.9288 43.8902 52.5648C43.8902 52.323 43.9378 52.0837 44.0303 51.8603C44.1229 51.637 44.2585 51.4341 44.4295 51.2632C44.6004 51.0922 44.8034 50.9567 45.0267 50.8642C45.2501 50.7718 45.4895 50.7242 45.7313 50.7243ZM26.8312 50.7243C27.1952 50.7244 27.551 50.8324 27.8536 51.0347C28.1562 51.237 28.392 51.5245 28.5312 51.8608C28.6705 52.1972 28.7068 52.5672 28.6358 52.9242C28.5647 53.2812 28.3893 53.6091 28.1319 53.8664C27.8745 54.1238 27.5466 54.299 27.1895 54.37C26.8325 54.4409 26.4625 54.4044 26.1262 54.2651C25.79 54.1258 25.5025 53.8899 25.3003 53.5873C25.0981 53.2846 24.9902 52.9288 24.9902 52.5648C24.9902 52.323 25.0378 52.0837 25.1303 51.8603C25.2229 51.637 25.3585 51.4341 25.5295 51.2632C25.7004 51.0922 25.9034 50.9567 26.1268 50.8642C26.3501 50.7718 26.5895 50.7242 26.8312 50.7243ZM26.8312 17.5947C27.1952 17.5948 27.551 17.7029 27.8536 17.9052C28.1562 18.1075 28.392 18.395 28.5312 18.7313C28.6705 19.0676 28.7068 19.4376 28.6358 19.7946C28.5647 20.1516 28.3893 20.4795 28.1319 20.7368C27.8745 20.9942 27.5466 21.1694 27.1895 21.2404C26.8325 21.3114 26.4625 21.2749 26.1262 21.1356C25.79 20.9962 25.5025 20.7603 25.3003 20.4577C25.0981 20.155 24.9902 19.7992 24.9902 19.4352C24.9902 19.1935 25.0378 18.9541 25.1303 18.7308C25.2229 18.5074 25.3585 18.3045 25.5295 18.1336C25.7004 17.9627 25.9034 17.8271 26.1268 17.7347C26.3501 17.6422 26.5895 17.5947 26.8312 17.5947ZM45.7313 17.5947C46.0952 17.5948 46.451 17.7029 46.7536 17.9052C47.0562 18.1075 47.292 18.395 47.4312 18.7313C47.5705 19.0676 47.6068 19.4376 47.5358 19.7946C47.4647 20.1516 47.2893 20.4795 47.0319 20.7368C46.7745 20.9942 46.4465 21.1694 46.0895 21.2404C45.7325 21.3114 45.3625 21.2749 45.0262 21.1356C44.69 20.9962 44.4025 20.7603 44.2003 20.4577C43.9981 20.155 43.8902 19.7992 43.8902 19.4352C43.8902 19.1935 43.9378 18.9541 44.0303 18.7308C44.1229 18.5074 44.2585 18.3045 44.4295 18.1336C44.6004 17.9627 44.8034 17.8271 45.0267 17.7347C45.2501 17.6422 45.4895 17.5947 45.7313 17.5947ZM55.3039 34.0369C55.6679 34.0369 56.0237 34.1449 56.3263 34.3472C56.629 34.5495 56.8648 34.8369 57.0041 35.1732C57.1433 35.5096 57.1797 35.8796 57.1087 36.2366C57.0376 36.5936 56.8623 36.9215 56.6049 37.1789C56.3475 37.4363 56.0196 37.6115 55.6626 37.6825C55.3055 37.7535 54.9355 37.7171 54.5992 37.5778C54.2629 37.4384 53.9755 37.2025 53.7733 36.8999C53.571 36.5972 53.4631 36.2414 53.4631 35.8774C53.4632 35.3892 53.6571 34.9211 54.0023 34.5759C54.3475 34.2308 54.8157 34.0369 55.3039 34.0369ZM17.2583 34.0369C17.6223 34.0369 17.9782 34.1449 18.2808 34.3472C18.5834 34.5495 18.8193 34.8369 18.9585 35.1732C19.0978 35.5096 19.1342 35.8796 19.0632 36.2366C18.9921 36.5936 18.8168 36.9215 18.5594 37.1789C18.302 37.4363 17.974 37.6115 17.617 37.6825C17.26 37.7535 16.89 37.7171 16.5537 37.5778C16.2174 37.4384 15.9299 37.2025 15.7277 36.8999C15.5255 36.5972 15.4176 36.2414 15.4176 35.8774C15.4176 35.6357 15.4652 35.3963 15.5577 35.173C15.6502 34.9497 15.7858 34.7468 15.9567 34.5758C16.1277 34.4049 16.3306 34.2694 16.5539 34.1769C16.7773 34.0844 17.0166 34.0368 17.2583 34.0369ZM15.9083 22.7481C16.1996 22.7481 16.4843 22.8344 16.7265 22.9963C16.9687 23.1581 17.1574 23.3881 17.2689 23.6572C17.3804 23.9263 17.4095 24.2224 17.3527 24.5081C17.2958 24.7937 17.1555 25.0561 16.9495 25.2621C16.7436 25.468 16.4811 25.6083 16.1955 25.6651C15.9098 25.7218 15.6137 25.6926 15.3446 25.5811C15.0755 25.4696 14.8455 25.2808 14.6838 25.0386C14.522 24.7964 14.4357 24.5117 14.4357 24.2204C14.4358 23.8299 14.591 23.4554 14.8671 23.1793C15.1433 22.9032 15.5178 22.7481 15.9083 22.7481ZM15.9083 46.307C16.1996 46.307 16.4843 46.3933 16.7265 46.5552C16.9687 46.717 17.1574 46.947 17.2689 47.2161C17.3804 47.4852 17.4095 47.7813 17.3527 48.067C17.2958 48.3527 17.1555 48.6151 16.9495 48.821C16.7436 49.0269 16.4811 49.1672 16.1955 49.224C15.9098 49.2807 15.6137 49.2515 15.3446 49.14C15.0755 49.0285 14.8455 48.8398 14.6838 48.5975C14.522 48.3553 14.4357 48.0706 14.4357 47.7793C14.4358 47.3888 14.591 47.0143 14.8671 46.7382C15.1433 46.4621 15.5178 46.307 15.9083 46.307ZM56.6539 46.307C56.9451 46.307 57.2299 46.3933 57.472 46.5552C57.7142 46.717 57.903 46.947 58.0144 47.2161C58.1259 47.4852 58.155 47.7813 58.0982 48.067C58.0413 48.3527 57.9011 48.6151 57.6951 48.821C57.4891 49.0269 57.2267 49.1672 56.941 49.224C56.6553 49.2807 56.3592 49.2515 56.0901 49.14C55.821 49.0285 55.5911 48.8398 55.4293 48.5975C55.2675 48.3553 55.1812 48.0706 55.1813 47.7793C55.1813 47.3888 55.3365 47.0143 55.6127 46.7382C55.8888 46.4621 56.2634 46.307 56.6539 46.307ZM56.6539 22.7481C56.9451 22.7481 57.2299 22.8344 57.472 22.9963C57.7142 23.1581 57.903 23.3881 58.0144 23.6572C58.1259 23.9263 58.155 24.2224 58.0982 24.5081C58.0413 24.7937 57.9011 25.0561 57.6951 25.2621C57.4891 25.468 57.2267 25.6083 56.941 25.6651C56.6553 25.7218 56.3592 25.6926 56.0901 25.5811C55.821 25.4696 55.5911 25.2808 55.4293 25.0386C55.2675 24.7964 55.1812 24.5117 55.1813 24.2204C55.1813 23.8299 55.3365 23.4554 55.6127 23.1793C55.8888 22.9032 56.2634 22.7481 56.6539 22.7481ZM36.2812 10.9688C36.5725 10.9688 36.8572 11.0552 37.0993 11.2171C37.3415 11.379 37.5302 11.609 37.6416 11.8781C37.753 12.1472 37.7821 12.4433 37.7252 12.7289C37.6684 13.0146 37.5281 13.277 37.3221 13.4829C37.1161 13.6888 36.8537 13.829 36.568 13.8858C36.2823 13.9425 35.9862 13.9133 35.7172 13.8018C35.4481 13.6903 35.2181 13.5015 35.0564 13.2593C34.8946 13.0171 34.8083 12.7323 34.8083 12.4411C34.8083 12.2477 34.8464 12.0562 34.9205 11.8775C34.9945 11.6989 35.103 11.5365 35.2398 11.3998C35.3766 11.2631 35.5389 11.1546 35.7176 11.0807C35.8963 11.0067 36.0879 10.9687 36.2812 10.9688ZM36.2812 58.0863C36.5725 58.0863 36.8572 58.1728 37.0993 58.3346C37.3415 58.4965 37.5302 58.7265 37.6416 58.9956C37.753 59.2646 37.7821 59.5607 37.7253 59.8464C37.6684 60.132 37.5281 60.3944 37.3222 60.6003C37.1162 60.8062 36.8538 60.9465 36.5682 61.0033C36.2825 61.0601 35.9864 61.0309 35.7174 60.9194C35.4483 60.8079 35.2183 60.6192 35.0565 60.377C34.8947 60.1349 34.8083 59.8502 34.8083 59.5589C34.8083 59.3655 34.8464 59.174 34.9205 58.9953C34.9945 58.8166 35.103 58.6543 35.2398 58.5175C35.3765 58.3808 35.5389 58.2723 35.7176 58.1983C35.8963 58.1243 36.0878 58.0862 36.2812 58.0863ZM49.2902 57.1047C49.5329 57.1047 49.7701 57.1767 49.9719 57.3115C50.1737 57.4464 50.331 57.638 50.4239 57.8622C50.5168 58.0864 50.5411 58.3332 50.4937 58.5712C50.4464 58.8092 50.3295 59.0279 50.1579 59.1995C49.9863 59.3711 49.7676 59.488 49.5296 59.5353C49.2916 59.5827 49.0448 59.5584 48.8206 59.4655C48.5964 59.3726 48.4047 59.2153 48.2699 59.0135C48.1351 58.8117 48.0631 58.5745 48.0631 58.3318C48.0631 58.1707 48.0948 58.0111 48.1565 57.8622C48.2182 57.7133 48.3086 57.5781 48.4225 57.4641C48.5364 57.3502 48.6717 57.2598 48.8206 57.1981C48.9695 57.1365 49.129 57.1047 49.2902 57.1047ZM23.272 57.1047C23.5147 57.1047 23.752 57.1767 23.9538 57.3115C24.1556 57.4464 24.3128 57.638 24.4057 57.8622C24.4986 58.0864 24.5229 58.3332 24.4755 58.5712C24.4282 58.8092 24.3113 59.0279 24.1397 59.1995C23.9681 59.3711 23.7495 59.488 23.5114 59.5353C23.2734 59.5827 23.0267 59.5584 22.8024 59.4655C22.5782 59.3726 22.3866 59.2153 22.2517 59.0135C22.1169 58.8117 22.0449 58.5745 22.0449 58.3318C22.0449 58.1707 22.0767 58.0111 22.1383 57.8622C22.2 57.7133 22.2904 57.5781 22.4043 57.4641C22.5183 57.3502 22.6536 57.2598 22.8024 57.1981C22.9513 57.1365 23.1109 57.1047 23.272 57.1047ZM23.272 12.4422C23.5147 12.4422 23.752 12.5142 23.9538 12.649C24.1556 12.7839 24.3128 12.9755 24.4057 13.1997C24.4986 13.424 24.5229 13.6707 24.4755 13.9087C24.4282 14.1467 24.3113 14.3654 24.1397 14.537C23.9681 14.7086 23.7495 14.8255 23.5114 14.8728C23.2734 14.9202 23.0267 14.8959 22.8024 14.803C22.5782 14.7101 22.3866 14.5528 22.2517 14.351C22.1169 14.1493 22.0449 13.912 22.0449 13.6693C22.0448 13.5081 22.0764 13.3484 22.138 13.1994C22.1996 13.0504 22.29 12.915 22.4039 12.8009C22.5179 12.6868 22.6532 12.5963 22.8022 12.5346C22.9511 12.4729 23.1108 12.4411 23.272 12.4411V12.4422ZM49.2902 12.4422C49.5329 12.4422 49.7701 12.5142 49.9719 12.649C50.1737 12.7839 50.331 12.9755 50.4239 13.1997C50.5168 13.424 50.5411 13.6707 50.4937 13.9087C50.4464 14.1467 50.3295 14.3654 50.1579 14.537C49.9863 14.7086 49.7676 14.8255 49.5296 14.8728C49.2916 14.9202 49.0448 14.8959 48.8206 14.803C48.5964 14.7101 48.4047 14.5528 48.2699 14.351C48.1351 14.1493 48.0631 13.912 48.0631 13.6693C48.0629 13.5081 48.0946 13.3484 48.1562 13.1994C48.2178 13.0504 48.3081 12.915 48.4221 12.8009C48.5361 12.6868 48.6714 12.5963 48.8203 12.5346C48.9693 12.4729 49.1289 12.4411 49.2902 12.4411V12.4422ZM62.0539 34.7735C62.2966 34.7735 62.5339 34.8455 62.7357 34.9803C62.9375 35.1152 63.0948 35.3069 63.1876 35.5311C63.2805 35.7554 63.3047 36.0021 63.2573 36.2402C63.21 36.4782 63.093 36.6969 62.9214 36.8685C62.7497 37.04 62.531 37.1569 62.2929 37.2042C62.0549 37.2514 61.8081 37.227 61.5839 37.1341C61.3597 37.0411 61.1681 36.8838 61.0333 36.6819C60.8985 36.48 60.8267 36.2427 60.8268 36C60.8268 35.8389 60.8585 35.6793 60.9202 35.5304C60.9819 35.3815 61.0722 35.2463 61.1862 35.1323C61.3001 35.0184 61.4354 34.928 61.5843 34.8663C61.7332 34.8046 61.8927 34.7729 62.0539 34.7729V34.7735ZM10.5083 34.7735C10.7511 34.7735 10.9883 34.8455 11.1901 34.9803C11.3919 35.1152 11.5492 35.3069 11.6421 35.5311C11.7349 35.7554 11.7592 36.0021 11.7118 36.2402C11.6644 36.4782 11.5475 36.6969 11.3758 36.8685C11.2042 37.04 10.9855 37.1569 10.7474 37.2042C10.5093 37.2514 10.2626 37.227 10.0384 37.1341C9.81415 37.0411 9.62255 36.8838 9.48778 36.6819C9.35301 36.48 9.28114 36.2427 9.28125 36C9.28125 35.8389 9.31299 35.6793 9.37466 35.5304C9.43632 35.3815 9.52671 35.2463 9.64066 35.1323C9.7546 35.0184 9.88988 34.928 10.0388 34.8663C10.1876 34.8046 10.3472 34.7729 10.5083 34.7729V34.7735Z"
-				fill={themes.colors.white}
+				fill={colors.white}
 			/>
 		</Svg>
 	);
