+++ src/types/IconType.ts
